rename node fs module
