+++ src/fs/importTs.ts
@@ -6,7 +6,7 @@ import {createEsbuildBuilder} from '../build/esbuildBuilder.js';
 import type {BuildConfig} from '../config/buildConfig.js';
 import type {BuildContext, TextBuildSource} from '../build/builder.js';
 import {DEFAULT_ECMA_SCRIPT_TARGET} from '../build/tsBuildHelpers.js';
-import {outputFile, readFile, remove} from './nodeFs.js';
+import {outputFile, readFile, remove} from './node.js';
 import {basename, dirname, join} from 'path';
 import {stripStart} from '../utils/string.js';
 import {isExternalNodeModule} from '../utils/module.js';
