+++ src/gen/genModule.ts
@@ -1,6 +1,6 @@
 import {ModuleMeta, loadModule, LoadModuleResult, findModules} from '../fs/modules.js';
 import {Gen, GenResults, GenFile, isGenPath, GEN_FILE_PATTERN} from './gen.js';
-import {pathExists, readFile, findFiles} from '../fs/nodeFs.js';
+import {pathExists, readFile, findFiles} from '../fs/node.js';
 import {getPossibleSourceIds} from '../fs/inputPath.js';
 import {paths} from '../paths.js';
 import type {Obj} from '../index.js';
