+++ src/project/packageJson.ts
@@ -1,6 +1,6 @@
 import {join} from 'path';
 
-import {readJson} from '../fs/nodeFs.js';
+import {readJson} from '../fs/node.js';
 import {paths, groPaths, isThisProjectGro} from '../paths.js';
 import type {Json} from '../utils/json.js';
 import type {Obj} from '../index.js';
